requirements.txt
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
flask==3.0.3
google-auth==2.34.0
google-auth-oauthlib==1.2.1
idna==3.8
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==2.1.5
oauthlib==3.2.2
pyasn1==0.6.1
pyasn1-modules==0.4.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
urllib3==2.2.3
werkzeug==3.0.4
zipp==3.20.1
